backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==1.0.3
python-dotenv==1.0.0
python-dateutil==2.8.2
pydantic==2.8.0
typing-extensions>=4.6.0
fastapi-mail==1.4.1
asyncpg==0.30.0
email-validator==2.1.0
openai==1.35.0
pgvector>=0.2.0
# Document Processing Libraries
PyPDF2==3.0.1
python-docx==1.1.0
